Crate evolution_wire

Wire data types for the evolution game.

Re-exports

pub use self::channel::Channel;

Modules

channel
TODO:

remote
Remote protocol wire datatypes.

Structs

Action4
The choice of a player's card choices for the round.

BT
A trading of a card for a new board.

Choice
The data a player gets to make a choice for the action step.

Configuration
A game state configuration.

Feeding
The information passed to a player for making a feed choice.

FoodValue
An integer in the range [-8, 8].

GB
A request to trade a card for a body size increase.

GP
A request to trade a card for a population increase.

Integer
An integer.

LOT
A list traits, which must be at most 3 in length.

Nat
A natural number in the range [0, 7].

NatPlus
A natural number in the range [1, 7].

Natural
Any integer >= 0.

NaturalPlus
Any integer > 0.

Player
A player possibly with owned data.

RT
The replacement of a sigle trait on a species.

Situation
An attacking situation between 2 to 4 species.

Species
A species with or without fat food.

SpeciesCard
A species card, with a food value, and a trait.

StartRound
The input to the xstep4 test harness.

Enums

Either
A wire type for containing one of two variants, without knowing the variant ahead of time.

FeedChoice
A wire feed choice response message.

Trait
A trait.

Traits

FromWire
A trait for getting meaningful data from wire data.

ToWire
A trait for getting wire data from meaningful data.

Type Definitions

LOC
A list of species cards.

LOP
A list of wire player pluses.

LOS
A list of species.

Step4
Represents all the choices of cards for a game's players.