Struct evolution_wire::channel::Channel

pub struct Channel(_, _);

Sending and receiving whole wire messages.

Methods

impl Channel

pub fn connect_to_socket_addr<A: ToSocketAddrs>(
info: String,
socket_addr: A
) -> Result<Channel, Error>

Create a stream, and connect to it.

pub fn connect_to_tcp_stream(
info: String,
stream: TcpStream
) -> Result<Channel, Error>

Connect to a tcp stream, we'll send the "info" for this channel, we must get back the response "ok".

pub fn accept_from_socket_addr<A: ToSocketAddrs>(
socket_addr: A
) -> Result<Channel, Error>

pub fn accept_from_tcp_stream(stream: TcpStream) -> Result<Channel, Error>

Accept from a tcp stream, we must get some "info" and return "ok". TODO: Boolean to tell channel it was not good?

pub fn info(&self) -> &str

pub fn call<D, C>(&mut self, domain: &D) -> Result<C, Error> where
D: Serialize + Debug,
C: Deserialize + Debug,

pub fn accept_call<D, C>(&mut self, func: &dyn Fn(D) -> C) -> Result<C, Error> where
D: Deserialize + Debug,
C: Serialize + Debug,

pub fn send<T: Serialize + Debug>(&mut self, message: &T) -> Result<(), Error>

pub fn recv<T: Deserialize + Debug>(&mut self) -> Result<T, Error>

pub fn recv_either<T, U>(&mut self) -> Result<Either<T, U>, Error> where
T: Deserialize + Debug,
U: Deserialize + Debug,

Receive an Either<T, U> type from the wire. This provides support for channel consumers to logically branch based on channel messages.