Module evolution_wire::remote

Remote protocol wire datatypes.

Specification

These datatypes were last updated from this link on 04/15/16.

Structs

Action4

Choose
A data type representing a request for a player choice during the Action step. The two lists of boards represent the players before the current, and after the current, respectively, in turn order.

GB

GP

Start
A wire data type used to signify the start of a new round.

State
A wire data type representing a given player's internal state knowledge as well as public knowledge of the rest of the game. Provides enough information to generate a choice during Feeding.

Type Definitions

Boards

Cards

FeedingChoice

LOB
The player’s complete current state and the public state of the competitors. Specifically, [bg,bd,c,w,others] consists of the player’s current bag, the current species boards, the current cards, the number of available food tokens, and the species boards of all other players (in turn order).

LOT

Species