Trait evolution_wire::ToWire

pub trait ToWire<W: Serialize> {
    fn to_wire(&self) -> W;
}

A trait for getting wire data from meaningful data.

Here the meaningful data is being converted into data for the wire, which should never fail. If the internal data set is larger than that of the wire, then there is a problem with the design of the wire data format that requires rethinking things.

Required methods

fn to_wire(&self) -> W

Loading content...